# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UAE Freelance Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: skill
          schema: { type: string }
        - in: query
          name: budgetType
          schema: { type: string }
        - in: query
          name: minBudget
          schema: { type: number }
        - in: query
          name: maxBudget
          schema: { type: number }
        - in: query
          name: mine
          schema: { type: boolean }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobCard"
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /jobs/{id}/complete:
    post:
      operationId: completeJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /saved-jobs/{jobId}:
    post:
      operationId: saveJob
      tags: [jobs]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Saved
    delete:
      operationId: unsaveJob
      tags: [jobs]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
...
  /payments/create-intent:
    post:
      operationId: createPaymentIntent
      tags: [payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentIntentBody"
      responses:
        "200":
          description: Intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentIntent"
...
components:
  schemas:
...
    BudgetType:
      type: string
      enum: [fixed, hourly]
    JobStatus:
      type: string
      enum: [open, in_progress, completed, closed]
    JobCard:
      type: object
      required: [id, title, categorySlug, categoryNameEn, budgetType, budgetMin, budgetMax, currency, skills, status, createdAt, proposalCount, clientName]
      properties:
        id: { type: integer }
        title: { type: string }
        descriptionShort: { type: string }
        categorySlug: { type: string }
        categoryNameEn: { type: string }
        categoryNameAr: { type: string }
        budgetType: { $ref: "#/components/schemas/BudgetType" }
        budgetMin: { type: number }
        budgetMax: { type: number }
        currency: { type: string }
        skills:
          type: array
          items: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        deadline: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        proposalCount: { type: integer }
        clientId: { type: integer }
        clientName: { type: string }
        clientCompany: { type: ["string", "null"] }
        clientLogoUrl: { type: ["string", "null"] }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/JobCard"
        - type: object
          required: [description]
          properties:
            description: { type: string }
            saved: { type: boolean }
    Job:
      type: object
      required: [id, title, description, categorySlug, budgetType, budgetMin, budgetMax, currency, skills, status, createdAt, clientId]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        categorySlug: { type: string }
        budgetType: { $ref: "#/components/schemas/BudgetType" }
        budgetMin: { type: number }
        budgetMax: { type: number }
        currency: { type: string }
        skills:
          type: array
          items: { type: string }
        status: { $ref: "#/components/schemas/JobStatus" }
        deadline: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        clientId: { type: integer }
    CreateJobBody:
      type: object
      required: [title, description, categorySlug, budgetType, budgetMin, budgetMax, skills]
      properties:
        title: { type: string }
        description: { type: string }
        categorySlug: { type: string }
        budgetType: { $ref: "#/components/schemas/BudgetType" }
        budgetMin: { type: number }
        budgetMax: { type: number }
        currency: { type: string }
        skills:
          type: array
          items: { type: string }
        deadline: { type: ["string", "null"], format: date-time }
    UpdateJobBody:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        categorySlug: { type: string }
        budgetType: { $ref: "#/components/schemas/BudgetType" }
        budgetMin: { type: number }
        budgetMax: { type: number }
        currency: { type: string }
        skills:
          type: array
          items: { type: string }
        deadline: { type: ["string", "null"], format: date-time }
        status: { $ref: "#/components/schemas/JobStatus" }
...
    CreatePaymentIntentBody:
      type: object
      required: [jobId, proposalId, amount]
      properties:
        jobId: { type: integer }
        proposalId: { type: integer }
        amount: { type: number }
        currency: { type: string }
    PaymentIntent:
      type: object
      required: [paymentId, clientSecret, amount, currency, status]
      properties:
        paymentId: { type: integer }
        clientSecret: { type: string }
        amount: { type: number }
        currency: { type: string }
        status: { type: string }
...