# lib/api-zod/src/generated/types/createPaymentIntentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAE Freelance Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePaymentIntentBody {
  jobId: number;
  proposalId: number;
  amount: number;
  currency?: string;
}