# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAE Freelance Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { JobCard } from "./jobCard";

export type JobDetail = JobCard & {
  description: string;
  saved?: boolean;
};