# lib/api-zod/src/generated/types/updateJobBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UAE Freelance Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetType } from "./budgetType";
import type { JobStatus } from "./jobStatus";

export interface UpdateJobBody {
  title?: string;
  description?: string;
  categorySlug?: string;
  budgetType?: BudgetType;
  budgetMin?: number;
  budgetMax?: number;
  currency?: string;
  skills?: string[];
  /** @nullable */
  deadline?: Date | null;
  status?: JobStatus;
}